Add gauge chart for accuracy and confidence
