+++ src/component/AccuracyInfo.jsx
@@ -1,7 +1,8 @@
-import React from "react";
+import React, { useLayoutEffect } from "react";
 import styled from "styled-components";
 
 import ModelAccuracy from "./ModelAccuracy";
+import ModelConfidence from "./GaugeChart";
 
 const Container = styled.div`
   display: flex;
@@ -20,15 +21,23 @@ const Confidence = styled.img`
 const Accuracy = styled.img`
   margin: 10px;
 `;
-const AboutModel = () => {
+const AccuracyInfo = ({ accuracy }) => {
   return (
     <Container>
       {/* <Accuracy src={`http://15.164.105.78:8000/accuracy}`} alt="model accuracy" /> */}
-      <p>모델의 정확도 : 80%</p>
-      <ModelAccuracy />
+      <p>AI의 정확도 : {accuracy}%</p>
+      <ModelConfidence
+        chartId="accuracy"
+        firstColor="#6ea6f4"
+        secondColor="#e8e8e8"
+        leftLabel=""
+        rightLabel=""
+        gaugeValue={80.3}
+      />
+      {/* <ModelAccuracy /> */}
 
       {/* <Confidence src={`15.164.105.78:8000/confidence/${pdi["id"]}`} alt="model confidence" /> */}
     </Container>
   );
 };
-export default AboutModel;
+export default AccuracyInfo;
